refactor: useGeolocation hook 로직 분리
